code-repo/gcp_data_protection_demo/functions/dlp_scan_trigger/requirements.txt
google-cloud-dlp
flask
google-cloud-bigquery